🌧 :: 동아리 확인 예외처리 추가

+++ src/main/java/com/example/daedongyeojido_server/domain/interview/application/ModifyInterviewTimeService.java
@@ -1,8 +1,10 @@
 package com.example.daedongyeojido_server.domain.interview.application;
 
+import com.example.daedongyeojido_server.domain.club.exception.ClubMisMatchException;
 import com.example.daedongyeojido_server.domain.interview.dao.InterviewTimeRepository;
 import com.example.daedongyeojido_server.domain.interview.domain.InterviewTime;
 import com.example.daedongyeojido_server.domain.interview.dto.request.InterviewTimeRequest;
+import com.example.daedongyeojido_server.domain.user.application.facade.UserFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,8 +17,13 @@ public class ModifyInterviewTimeService {
 
     private final InterviewTimeRepository interviewTimeRepository;
 
+    private final UserFacade userFacade;
+
     @Transactional
     public void modifyInterviewTime(String clubName, List<InterviewTimeRequest> requests) {
+        if (!(userFacade.currentUser().getMyClub().getClubName().equals(clubName)))
+            throw ClubMisMatchException.EXCEPTION;
+
         interviewTimeRepository.deleteAllByClubName(clubName);
 
         for (int i = 0; i < requests.size(); i++) {
